Abstract route logic into ShortlinkController class
